+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - added new `:vectorized` instructions to `:vector` as appropriate
 - changed `:x-print` instruction to return the string result of `pr-str`
 - modified PushDSL `:needs` and `:products` table to remove 0 values; since it was originally written, stacks are created on demand
+- cleaned up `push.type.item.vector` and `push.type.item.vectorized` call structures for legibility
 
 ## 0.1.21